Feat: Expanded npc->HitBy[] Changelog: npc->HitBy[] is now size 16, and records the weapon type that struck an npc, and the parent item that generated that weapon.

+++ src/guys.h
@@ -80,7 +80,7 @@ public:
     byte defense[edefLAST255];
     byte hitsfx,deadsfx;
     
-    byte hitby[NUM_HIT_TYPES_USED];
+    long hitby[NUM_HIT_TYPES_USED];
     short firesfx;
     bool isCore;
     short parentCore;
